feat: use ReVanced base package name for GmsCore
