+++ src/main/kotlin/app/revanced/patches/music/misc/gms/fingerprints/CastContextFetchFingerprint.kt
@@ -1,0 +1,8 @@
+package app.revanced.patches.music.misc.gms.fingerprints
+
+
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+
+object CastContextFetchFingerprint : MethodFingerprint(
+    strings = listOf("Error fetching CastContext.")
+)
