+++ src/main/kotlin/app/revanced/patches/music/misc/gms/fingerprints/GooglePlayUtilityFingerprint.kt
@@ -1,0 +1,18 @@
+package app.revanced.patches.music.misc.gms.fingerprints
+
+import app.revanced.patcher.extensions.or
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+import org.jf.dexlib2.AccessFlags
+
+object GooglePlayUtilityFingerprint : MethodFingerprint(
+    "I",
+    AccessFlags.PUBLIC or AccessFlags.STATIC,
+    listOf("L", "I"),
+    strings = listOf(
+        "This should never happen.",
+        "MetadataValueReader",
+        "GooglePlayServicesUtil",
+        "com.android.vending",
+        "android.hardware.type.embedded"
+    )
+)
