+++ src/main/kotlin/app/revanced/patches/music/misc/gms/shared/Constants.kt
@@ -1,0 +1,9 @@
+package app.revanced.patches.music.misc.gms.shared
+
+object Constants {
+    internal const val REVANCED_MUSIC_APP_NAME = "YT Music ReVanced"
+    internal const val REVANCED_MUSIC_PACKAGE_NAME = "app.revanced.android.apps.youtube.music"
+    internal const val MUSIC_PACKAGE_NAME = "com.google.android.apps.youtube.music"
+    internal const val SPOOFED_PACKAGE_NAME = MUSIC_PACKAGE_NAME
+    internal const val SPOOFED_PACKAGE_SIGNATURE = "afb0fed5eeaebdd86f56a97742f4b6b33ef59875"
+}
