--- src/main/kotlin/app/revanced/patches/music/misc/microg/fingerprints/CastContextFetchFingerprint.kt
@@ -1,8 +1,0 @@
-package app.revanced.patches.music.misc.microg.fingerprints
-
-
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-
-object CastContextFetchFingerprint : MethodFingerprint(
-    strings = listOf("Error fetching CastContext.")
-)
