+++ src/main/kotlin/app/revanced/patches/youtube/misc/gms/annotations/GmsCoreCompatibility.kt
@@ -1,0 +1,8 @@
+package app.revanced.patches.youtube.misc.gms.annotations
+
+import app.revanced.patcher.annotation.Compatibility
+import app.revanced.patcher.annotation.Package
+
+@Compatibility([Package("com.google.android.youtube", arrayOf("18.16.37", "18.19.35", "18.20.39", "18.23.35"))])
+@Target(AnnotationTarget.CLASS)
+internal annotation class GmsCoreCompatibility
