+++ src/main/kotlin/app/revanced/patches/youtube/misc/gms/fingerprints/CastDynamiteModuleFingerprint.kt
@@ -1,0 +1,8 @@
+package app.revanced.patches.youtube.misc.gms.fingerprints
+
+
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+
+object CastDynamiteModuleFingerprint : MethodFingerprint(
+    strings = listOf("com.google.android.gms.cast.framework.internal.CastDynamiteModuleImpl")
+)
