+++ src/main/kotlin/app/revanced/patches/youtube/misc/gms/fingerprints/GooglePlayUtilityFingerprint.kt
@@ -1,0 +1,12 @@
+package app.revanced.patches.youtube.misc.gms.fingerprints
+
+import app.revanced.patcher.extensions.or
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+import org.jf.dexlib2.AccessFlags
+
+object GooglePlayUtilityFingerprint : MethodFingerprint(
+    returnType = "I",
+    accessFlags = AccessFlags.PUBLIC or AccessFlags.STATIC,
+    parameters = listOf("L", "I"),
+    strings = listOf("This should never happen.", "MetadataValueReader", "com.google.android.gms")
+)
