+++ src/main/kotlin/app/revanced/patches/youtube/misc/gms/fingerprints/ServiceCheckFingerprint.kt
@@ -1,0 +1,12 @@
+package app.revanced.patches.youtube.misc.gms.fingerprints
+
+import app.revanced.patcher.extensions.or
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
+import org.jf.dexlib2.AccessFlags
+
+object ServiceCheckFingerprint : MethodFingerprint(
+    returnType = "V",
+    accessFlags = AccessFlags.PUBLIC or AccessFlags.STATIC,
+    parameters = listOf("L", "I"),
+    strings = listOf("Google Play Services not available", "GooglePlayServices not available due to error ")
+)
