--- src/main/kotlin/app/revanced/patches/youtube/misc/microg/annotations/MicroGPatchCompatibility.kt
@@ -1,8 +1,0 @@
-package app.revanced.patches.youtube.misc.microg.annotations
-
-import app.revanced.patcher.annotation.Compatibility
-import app.revanced.patcher.annotation.Package
-
-@Compatibility([Package("com.google.android.youtube", arrayOf("18.16.37", "18.19.35", "18.20.39", "18.23.35"))])
-@Target(AnnotationTarget.CLASS)
-internal annotation class MicroGPatchCompatibility
