--- src/main/kotlin/app/revanced/patches/youtube/misc/microg/fingerprints/CastDynamiteModuleFingerprint.kt
@@ -1,8 +1,0 @@
-package app.revanced.patches.youtube.misc.microg.fingerprints
-
-
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-
-object CastDynamiteModuleFingerprint : MethodFingerprint(
-    strings = listOf("com.google.android.gms.cast.framework.internal.CastDynamiteModuleImpl")
-)
