--- src/main/kotlin/app/revanced/patches/youtube/misc/microg/fingerprints/CastDynamiteModuleV2Fingerprint.kt
@@ -1,8 +1,0 @@
-package app.revanced.patches.youtube.misc.microg.fingerprints
-
-
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-
-object CastDynamiteModuleV2Fingerprint : MethodFingerprint(
-    strings = listOf("Failed to load module via V2: ")
-)
