--- src/main/kotlin/app/revanced/patches/youtube/misc/microg/fingerprints/PrimeFingerprint.kt
@@ -1,8 +1,0 @@
-package app.revanced.patches.youtube.misc.microg.fingerprints
-
-
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-
-object PrimeFingerprint : MethodFingerprint(
-    strings = listOf("com.google.android.GoogleCamera", "com.android.vending")
-)
