--- src/main/kotlin/app/revanced/patches/youtube/misc/microg/fingerprints/ServiceCheckFingerprint.kt
@@ -1,12 +1,0 @@
-package app.revanced.patches.youtube.misc.microg.fingerprints
-
-import app.revanced.patcher.extensions.or
-import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-import org.jf.dexlib2.AccessFlags
-
-object ServiceCheckFingerprint : MethodFingerprint(
-    returnType = "V",
-    accessFlags = AccessFlags.PUBLIC or AccessFlags.STATIC,
-    parameters = listOf("L", "I"),
-    strings = listOf("Google Play Services not available", "GooglePlayServices not available due to error ")
-)
